+++ datamigratie.client/src/components/AlertInline.vue
@@ -1,9 +1,11 @@
 <template>
-  <section class="notice" aria-live="polite" role="alert">
+  <section class="notice" :class="type && `notice--${type}`" aria-live="polite" role="alert">
     <slot></slot>
   </section>
 </template>
 
-<script setup lang="ts"></script>
+<script setup lang="ts">
+defineProps<{ type?: "warning" | "info" }>();
+</script>
 
 <style lang="scss" scoped></style>