+++ docs/changelog/changelog.md
@@ -17,8 +17,10 @@
 - [UI changes: starting a migration](https://dimpact.atlassian.net/browse/DATA-408)
 - [Migrate Betrokkenen (Rollen-related)](https://dimpact.atlassian.net/browse/DATA-333)
 - [Migrate Initiator (Rollen-related)](https://dimpact.atlassian.net/browse/DATA-167)
+- [Automatic refresh of migration status](https://dimpact.atlassian.net/browse/DATA-403)
 - [Limit error message size](https://dimpact.atlassian.net/browse/DATA-339)
 - [Parallel zaak migration](https://dimpact.atlassian.net/browse/DATA-617)
+- [Make error message expandable in UI](https://dimpact.atlassian.net/browse/DATA-379)
 
 ## v0.1.0